feat(travis): add travis integration
