+++ .travis.yml
@@ -1,0 +1,7 @@
+language: node_js
+node_js:
+	- "6"
+	- "8"
+	- "10"
+script:
+	- npm run test:ci
