+++ package.json
@@ -17,22 +17,21 @@
     "bugs": "https://github.com/TinkoffCreditSystems/angular-contenteditable-accessor/issues",
     "homepage": "https://github.com/TinkoffCreditSystems/angular-contenteditable-accessor#README",
     "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build && npm run package",
-        "test": "ng test --browsers ChromeHeadless",
-        "lint": "ng lint",
-        "release": "git add . && standard-version -a",
-        "package": "node scripts/package.js"
-    },
+		"ng": "ng",
+		"start": "ng serve",
+		"build": "ng build && npm run postbuild",
+		"lint": "ng lint",
+		"test": "ng test --browsers ChromeHeadless --watch=false --code-coverage",
+		"test:ci": "npm run test && cat coverage/angular-contenteditable-accessor/lcov.info | coveralls",
+		"release": "standard-version && npm run build",
+		"package": "node scripts/updatePackage.js && git add projects/angular-contenteditable-accessor/package.json",
+		"postbuild": "node scripts/copyReadme.js && node scripts/removeDependencies.js",
+		"check": "npm run lint && tsc --noEmit --skipLibCheck && npm run test && lint-staged",
+		"precommit": "npm run check && npm run package"
+	},
     "husky": {
         "hooks": {
-            "pre-commit": "tsc --noEmit --skipLibCheck && npm run test && npm run build && git add . && lint-staged"
-        }
-    },
-    "standard-version": {
-        "scripts": {
-            "postbump": "npm run package && git add ."
+            "pre-commit": "npm run precommit"
         }
     },
     "lint-staged": {
@@ -65,6 +64,7 @@
         "@tinkoff/linters": "^0.3.0",
         "@types/jasmine": "~2.8.8",
         "@types/jasminewd2": "~2.0.3",
+        "coveralls": "^3.0.3",
         "husky": "^1.3.1",
         "jasmine-core": "~2.99.1",
         "jasmine-spec-reporter": "~4.2.1",
